Introduce the Query type for TemplateVar.Query ambiguity
In some cases, when importing dashboards that have Prometheus
datasources the `TemplateVar.Query` property may contain a nested object
instead of the plain query directly as a string.

Although it seems that Grafana is able to handle both cases when
importing the Unmarshal process fails. The original (deprecated) API
treats the entire dashboard as a `map[string]interface{}` which is not
very helpful for our case.

This commit adds a new `Query` type that will unmarshal the raw string
attribute into the `Query.Query` property. When marshalling the struct
if there is no `Query.RefId` then it will generate a string literal,
otherwise the nested object will be produced.