Add support for folder permissions
Some types to make it easier to interact with the folder permissions
were added.

`FolderPermissionItem` only a pair of [Role|TeamID|UserID] and
Permission is supported (i.e Role, TeamID, UserID) are mutually
exclusive but we do not validate it in the API code. For this case, the
validation is delegated to the Grafana server.

The following constants:

UserViewer
UserEditor
UserAdmin

were added to `user.go`.