Add ToItem helper method
The `ToItem` helper function translates a `FolderPermission` object into
its equivalent `FolderPermissionItem` one. This is quite often needed
when we want to add a new permission to a folder. The API expects both
the old permissions plus the new one otherwise the ones not included in
the request will be removed.

Signed-off-by: Jorge Luis Betancourt Gonzalez <jorge-luis.betancourt@trivago.com>